(GH-100) Refactor the Extension for common simple interfaces
The Debug Adapter requires similar object and configuration options from the
extension.  This commit refactors out the simple interfaces (i.e. cannot include
complex types, particularly from the vscode namespace) so that it can be
consumed by an out-of-process DebugAdapter.
