+++ client/src/interfaces.ts
@@ -1,0 +1,27 @@
+'use strict';
+
+// Only add simple interfaces here. No import's allowed
+
+export enum ConnectionStatus {
+  NotStarted,
+  Starting,
+  Running,
+  Stopping,
+  Failed
+}
+
+export enum ConnectionType {
+  Unknown,
+  Local,
+  Remote
+}
+
+export interface IConnectionConfiguration {
+  type: ConnectionType;
+  host: string;
+  port: number;
+  timeout: number;
+  preLoadPuppet: boolean;
+  debugFilePath: string;
+  puppetAgentDir: string;
+}
